reordered configs and utils in separate modules to be accessible
