+++ assignment_2/utils/data.py
@@ -1,8 +1,10 @@
 import os
-from pyspark.sql import SparkSession
-from pyspark.sql import DataFrame
+from pyspark.sql import SparkSession, DataFrame
+from pyspark.sql.types import DataType, StringType, IntegerType, FloatType, DateType, DecimalType, BooleanType
+from pyspark.sql.functions import regexp_extract, col
 
-from helpers_data_processing import validate_file_path
+from .validators import validate_file_path
+
 
 def load_data(spark: SparkSession, input_directory: str, partition: str | None) -> DataFrame:
     """
@@ -57,3 +59,43 @@ def load_data(spark: SparkSession, input_directory: str, partition: str | None) 
         
         print(f"Loaded data from all files from {input_directory}. Row count: {df.count()}")
         return df
+
+
+def transform_data_in_column(df: DataFrame, column_name: str, dtype: DataType):
+    """
+    Transforms data from a specified column in a DataFrame.
+    Args:
+        df (DataFrame): The DataFrame to transform data in.
+        column_name (str): The name of the column to transform the data in.
+        dtype (DataType): The target data type to transform the column to.
+    Returns:
+        df (DataFrame): The DataFrame with the transformed column.
+    """
+
+    # Regex pattern depending on the data type
+    if isinstance(dtype, StringType):
+        pattern = r'^[a-zA-Z0-9_]+$'  # Alphanumeric and underscore
+    elif isinstance(dtype, IntegerType):
+        pattern = r'^\d+$'  # Digits only
+    elif isinstance(dtype, FloatType) or isinstance(dtype, DecimalType):
+        pattern = r'^\d+(\.\d+)?$'  # Digits with optional decimal point
+    elif isinstance(dtype, BooleanType):
+        pattern = r'^(true|false|1|0)$'
+    elif isinstance(dtype, DateType):
+        pattern = r'^\d{4}-\d{2}-\d{2}$'  # YYYY-MM-DD format
+    else:
+        raise ValueError(f"Unsupported data type: {dtype}")
+    
+    # Check if the column exists in the DataFrame
+    if column_name not in df.columns:
+        raise ValueError(f"Column '{column_name}' does not exist in the DataFrame.")
+    
+    # Extract the matching part of the string using regexp_extract
+    df = df.withColumn(
+        column_name,
+        regexp_extract(col(column_name).cast(StringType()), pattern, 0)
+    )
+
+    # Cast the column to the specified data type & return the DataFrame
+    return df.withColumn(column_name, df[column_name].try_cast(dtype))
+
