+++ assignment_2/utils/validators.py
@@ -3,8 +3,7 @@ import re
 from datetime import datetime
 
 from pyspark.sql import DataFrame
-from pyspark.sql.types import DataType, StringType, IntegerType, FloatType, DateType, DecimalType, BooleanType
-from pyspark.sql.functions import regexp_extract, col, count
+from pyspark.sql.functions import count
 
 
 def validate_date(date_str: str, output_DateType: bool = False) -> datetime | str:
@@ -51,46 +50,7 @@ def build_partition_name(level: str, table_name: str, date: str, type: str) -> s
     return f"{level}_{table_name}_{date.replace('-', '_')}.{type}"
 
 
-def transform_data_in_column(df: DataFrame, column_name: str, dtype: DataType):
-    """
-    Transforms data from a specified column in a DataFrame.
-    Args:
-        df (DataFrame): The DataFrame to transform data in.
-        column_name (str): The name of the column to transform the data in.
-        dtype (DataType): The target data type to transform the column to.
-    Returns:
-        df (DataFrame): The DataFrame with the transformed column.
-    """
-
-    # Regex pattern depending on the data type
-    if isinstance(dtype, StringType):
-        pattern = r'^[a-zA-Z0-9_]+$'  # Alphanumeric and underscore
-    elif isinstance(dtype, IntegerType):
-        pattern = r'^\d+$'  # Digits only
-    elif isinstance(dtype, FloatType) or isinstance(dtype, DecimalType):
-        pattern = r'^\d+(\.\d+)?$'  # Digits with optional decimal point
-    elif isinstance(dtype, BooleanType):
-        pattern = r'^(true|false|1|0)$'
-    elif isinstance(dtype, DateType):
-        pattern = r'^\d{4}-\d{2}-\d{2}$'  # YYYY-MM-DD format
-    else:
-        raise ValueError(f"Unsupported data type: {dtype}")
-    
-    # Check if the column exists in the DataFrame
-    if column_name not in df.columns:
-        raise ValueError(f"Column '{column_name}' does not exist in the DataFrame.")
-    
-    # Extract the matching part of the string using regexp_extract
-    df = df.withColumn(
-        column_name,
-        regexp_extract(col(column_name).cast(StringType()), pattern, 0)
-    )
-
-    # Cast the column to the specified data type & return the DataFrame
-    return df.withColumn(column_name, df[column_name].try_cast(dtype))
-
-
-def pyspark_df_info(df: DataFrame):
+def pyspark_info(df: DataFrame):
     """
     Prints information about a PySpark DataFrame, similar to pandas.DataFrame.info().
     Includes total number of rows, and for each column: its name,
